Chore: move files under question-list directory part of SWM-164 related to #67
